refactor remote test running
